[CEN-1363] Refactor tests
