+++ src/test/java/it/gov/pagopa/rtd/ms/rtdmsdecrypter/model/BlobApplicationAwareTest.java
@@ -1,0 +1,25 @@
+package it.gov.pagopa.rtd.ms.rtdmsdecrypter.model;
+
+import static org.junit.jupiter.api.Assertions.assertSame;
+
+import org.junit.jupiter.api.Test;
+
+public class BlobApplicationAwareTest {
+  @Test
+  void shouldMatchRegex() {
+    String container = "rtd-transactions-32489876908u74bh781e2db57k098c5ad034341i8u7y";
+    String blob = "CSTAR.99910.TRNLOG.20220228.103107.001.csv.pgp";
+    String blobUri = "/blobServices/default/containers/" + container + "/blobs/" + blob;
+    BlobApplicationAware myBlob = new BlobApplicationAware(blobUri);
+    assertSame(BlobApplicationAware.Application.RTD, myBlob.getApp() );
+  }
+
+  @Test
+  void shouldNotMatchRegex() {
+    String container = "rtd-loremipsum-32489876908u74bh781e2db57k098c5ad034341i8u7y";
+    String blob = "CSTAR.99910.TRNLOG.20220228.103107.001.csv.pgp";
+    String blobUri = "/blobServices/default/containers/" + container + "/blobs/" + blob;
+    BlobApplicationAware myBlob = new BlobApplicationAware(blobUri);
+    assertSame(BlobApplicationAware.Application.NOAPP, myBlob.getApp() );
+  }
+}
